# Conflicts:
#	pom.xml
#	server/backend/Dockerfile
#	server/docker-compose.yml
#	src/main/java/kz/mb/project/mb_project/controller/UserController.java
#	src/main/java/kz/mb/project/mb_project/converter/CreateUserRequestToKUserConverter.java
#	src/main/java/kz/mb/project/mb_project/dto/CreateUserRequest.java
#	src/main/java/kz/mb/project/mb_project/dto/LoginResponseDto.java
#	src/main/java/kz/mb/project/mb_project/dto/SmsResponse.java
#	src/main/java/kz/mb/project/mb_project/dto/SuccessMessage.java
#	src/main/java/kz/mb/project/mb_project/dto/TokenResponse.java
#	src/main/java/kz/mb/project/mb_project/dto/UserLockedResponse.java
#	src/main/java/kz/mb/project/mb_project/dto/keycloak/KAccess.java
#	src/main/java/kz/mb/project/mb_project/dto/keycloak/KUser.java
#	src/main/java/kz/mb/project/mb_project/entity/AbstractEntity.java
#	src/main/java/kz/mb/project/mb_project/entity/Business.java
#	src/main/java/kz/mb/project/mb_project/entity/BusinessType.java
#	src/main/java/kz/mb/project/mb_project/entity/Otp.java
#	src/main/java/kz/mb/project/mb_project/entity/UserBusiness.java
#	src/main/java/kz/mb/project/mb_project/entity/UserDetail.java
#	src/main/java/kz/mb/project/mb_project/exception/ErrorMessage.java
#	src/main/java/kz/mb/project/mb_project/exception/GlobalExceptionHandler.java
#	src/main/java/kz/mb/project/mb_project/repo/BusinessRepository.java
#	src/main/java/kz/mb/project/mb_project/repo/UserBusinessRepository.java
#	src/main/java/kz/mb/project/mb_project/repo/UsersRepository.java
#	src/main/java/kz/mb/project/mb_project/scheduler/UserTask.java
#	src/main/java/kz/mb/project/mb_project/service/KeycloakServiceImpl.java
#	src/main/java/kz/mb/project/mb_project/service/SmsService.java
#	src/main/java/kz/mb/project/mb_project/service/UserService.java
#	src/main/java/kz/mb/project/mb_project/service/UserServiceImpl.java
#	src/main/java/kz/mb/project/mb_project/utils/PhoneNumberUtils.java
#	src/main/resources/application-dev.yml
#	src/main/resources/application-server.yml
#	src/main/resources/application.yaml
#	src/main/resources/db/changelog/DDL/001_init.sql
#	src/main/resources/db/changelog/db.changelog-master.yaml

+++ src/main/java/kz/mb/project/mb_project/service/UserServiceImpl.java
@@ -4,25 +4,30 @@ import java.util.Arrays;
 import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
+import java.util.function.Function;
+
 import lombok.extern.slf4j.Slf4j;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.core.convert.ConversionService;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+
 import kz.mb.project.mb_project.converter.KUserLoginResponseConverter;
 import kz.mb.project.mb_project.dto.CreateUserRequest;
 import kz.mb.project.mb_project.dto.LoginRequest;
 import kz.mb.project.mb_project.dto.LoginResponseDto;
 import kz.mb.project.mb_project.dto.OtpDto;
-import kz.mb.project.mb_project.dto.SmsResponse;
 import kz.mb.project.mb_project.dto.TokenResponse;
+import kz.mb.project.mb_project.dto.UpdateUserRequest;
 import kz.mb.project.mb_project.dto.UserLockedResponse;
 import kz.mb.project.mb_project.dto.keycloak.CreateKUser;
 import kz.mb.project.mb_project.dto.keycloak.KAction;
 import kz.mb.project.mb_project.dto.keycloak.KUser;
 import kz.mb.project.mb_project.entity.UserBusiness;
 import kz.mb.project.mb_project.entity.UserDetail;
+import kz.mb.project.mb_project.entity.UserRole;
 import kz.mb.project.mb_project.exception.ErrorMessage;
 import kz.mb.project.mb_project.exception.ForbiddenException;
 import kz.mb.project.mb_project.exception.FoundException;
@@ -30,9 +35,11 @@ import kz.mb.project.mb_project.exception.InternalServerException;
 import kz.mb.project.mb_project.exception.InvalidRequestException;
 import kz.mb.project.mb_project.exception.NotAuthorizedException;
 import kz.mb.project.mb_project.exception.NotFoundException;
+import kz.mb.project.mb_project.repo.BusinessRepository;
 import kz.mb.project.mb_project.repo.UserBusinessRepository;
 import kz.mb.project.mb_project.repo.UsersRepository;
 import kz.mb.project.mb_project.utils.PhoneNumberUtils;
+import kz.mb.project.mb_project.utils.ValidationUtils;
 
 @Service
 @Slf4j
@@ -43,8 +50,16 @@ public class UserServiceImpl implements UserService {
   protected final KeycloakService keycloakService;
   protected final KUserLoginResponseConverter kUserLoginResponseConverter;
   protected final UserBusinessRepository userBusinessRepository;
-  protected final SmsService smsService;
+
+  protected final NotificationService notificationService;
+
   protected final OtpService otpService;
+  protected final BusinessRepository businessRepository;
+
+  static Function<String, String> getIDFromUrl = url -> {
+    String[] uidParts = url.split("/");
+    return uidParts[uidParts.length - 1];
+  };
 
   public UserServiceImpl(
       @Autowired
@@ -54,38 +69,83 @@ public class UserServiceImpl implements UserService {
       ConversionService conversionService,
       KeycloakService keycloakService,
       KUserLoginResponseConverter kUserLoginResponseConverter,
-      UserBusinessRepository userBusinessRepository, SmsService smsService, OtpService otpService) {
+      UserBusinessRepository userBusinessRepository,
+      NotificationService notificationService, OtpService otpService,
+      BusinessRepository businessRepository) {
     this.usersRepository = usersRepository;
     this.conversionService = conversionService;
     this.keycloakService = keycloakService;
     this.kUserLoginResponseConverter = kUserLoginResponseConverter;
     this.userBusinessRepository = userBusinessRepository;
-    this.smsService = smsService;
+    this.notificationService = notificationService;
     this.otpService = otpService;
+    this.businessRepository = businessRepository;
   }
 
   @Override
   @Transactional
   public void createUser(CreateUserRequest createUserRequest) {
-    usersRepository.findUserDetailByUsername(createUserRequest.getPhone_number())
-        .ifPresent(userDetail -> {
-          throw new FoundException(ErrorMessage.USER_FOUND_EXCEPTION);
-        });
     TokenResponse token = keycloakService.getClientCredentialToken().block();
+
     if (token == null) {
       throw new NotAuthorizedException(ErrorMessage.AUTHORIZATION_ERROR);
     }
+
+    if (usersRepository.findUserDetailByUsername(createUserRequest.getPhone_number()).isPresent()
+        && keycloakService.getUsers(token).stream()
+        .anyMatch(user -> user.getUsername().equals(createUserRequest.getPhone_number()))) {
+      throw new FoundException(ErrorMessage.USER_FOUND_EXCEPTION);
+    }
+
+    if (PhoneNumberUtils.ensureKzCtnWithCountryCode(createUserRequest.getPhone_number()) == null) {
+      throw new InvalidRequestException(ErrorMessage.CREATE_INCORRECT_PHONE_NUMBER);
+    }
+    if (usersRepository.findUserDetailByEmail(createUserRequest.getEmail()).isPresent()
+        && keycloakService.getUsers(token).stream()
+        .anyMatch(user -> user.getEmail().equals(createUserRequest.getEmail()))) {
+      throw new FoundException(ErrorMessage.USER_CREATE_EMAIL_EXCEPTION);
+    }
+
+    if (!ValidationUtils.validateEmail(createUserRequest.getEmail())) {
+      throw new InvalidRequestException(ErrorMessage.INVALID_EMAIL);
+    }
+
     CreateKUser user = conversionService.convert(createUserRequest, CreateKUser.class);
     String uid = keycloakService.createUser(user, token);
+    uid = getIDFromUrl.apply(uid);
     if (uid == null) {
       throw new InternalServerException(ErrorMessage.USER_CREATE_EXCEPTION);
     }
-    String[] uidParts = uid.split("/");
-    uid = uidParts[uidParts.length - 1];
+
     UserDetail detail = UserDetail.builder().username(createUserRequest.getPhone_number())
-        .id(UUID.fromString(uid)).temporal(true).build();
+        .id(UUID.fromString(uid)).temporal(true).firstName(createUserRequest.getFirstname())
+        .lastName(createUserRequest.getLastname())
+        .email(createUserRequest.getEmail()).build();
     usersRepository.save(detail);
   }
+
+  @Override
+  public void updateUser(UpdateUserRequest updateUserRequest) {
+    TokenResponse token = keycloakService.getClientCredentialToken().block();
+
+    if (token == null) {
+      throw new NotAuthorizedException(ErrorMessage.AUTHORIZATION_ERROR);
+    }
+    if (usersRepository.findUserDetailByUsername(updateUserRequest.getPhone_number()).isEmpty()
+        && keycloakService.getUsers(token).stream()
+        .noneMatch(user -> user.getUsername().equals(updateUserRequest.getPhone_number()))) {
+      throw new FoundException(ErrorMessage.USER_NOT_FOUND_EXCEPTION);
+    }
+    KUser user = conversionService.convert(updateUserRequest, KUser.class);
+    keycloakService.updateUser(user,token);
+    UserDetail detail = UserDetail.builder().username(updateUserRequest.getPhone_number())
+        .id(updateUserRequest.getId()).temporal(updateUserRequest.getToTemporal())
+        .firstName(updateUserRequest.getFirstname())
+        .lastName(updateUserRequest.getLastname())
+        .email(updateUserRequest.getEmail()).build();
+    usersRepository.save(detail);
+  }
+
 
   @Override
   @Transactional
@@ -99,18 +159,20 @@ public class UserServiceImpl implements UserService {
     if (token == null) {
       throw new NotAuthorizedException(ErrorMessage.AUTHORIZATION_ERROR);
     }
-    UserLockedResponse locked = keycloakService.isUserLocked(user.get().getId().toString(), token).block();
-    if(locked != null && locked.getDisabled()){
+    UserLockedResponse locked = keycloakService.isUserLocked(user.get().getId().toString(), token)
+        .block();
+    if (locked != null && locked.getDisabled()) {
       KUser kUser = KUser.builder().id(user.get().getId().toString())
-          .requiredActions(new KAction[]{KAction.UPDATE_PASSWORD, KAction.CONFIGURE_TOTP}).totp(false).build();
+          .requiredActions(new KAction[]{KAction.UPDATE_PASSWORD, KAction.CONFIGURE_TOTP})
+          .totp(false).build();
       keycloakService.updateUser(kUser, token);
       throw new ForbiddenException(ErrorMessage.USER_LOCKED);
     }
     return keycloakService.getToken(loginRequest.getUsername(), loginRequest.getPassword()).block();
   }
 
+
   @Override
-  @Transactional
   public void setPassword(String username, String password) {
     Optional<UserDetail> user = usersRepository.findUserDetailByUsername(
         username);
@@ -125,13 +187,14 @@ public class UserServiceImpl implements UserService {
     if (kUser == null) {
       throw new NotFoundException(ErrorMessage.USER_NOT_FOUND_EXCEPTION);
     }
-    if(Arrays.asList(kUser.getRequiredActions()).contains(KAction.CONFIGURE_TOTP)){
+    if (Arrays.asList(kUser.getRequiredActions()).contains(KAction.CONFIGURE_TOTP)) {
       throw new InvalidRequestException(ErrorMessage.USER_UPDATE_EXCEPTION);
     }
     kUser.setRequiredActions(new KAction[]{});
     UserDetail toUpdateUser = user.get();
     toUpdateUser.setTemporal(false);
     kUser.setEnabled(true);
+
     keycloakService.setCredentials(kUser.getId(), password, token);
     keycloakService.updateUser(kUser, token);
     usersRepository.save(toUpdateUser);
@@ -154,8 +217,10 @@ public class UserServiceImpl implements UserService {
     }
     List<UserBusiness> userBusiness = userBusinessRepository.findAllByUserUsername(username);
     LoginResponseDto loginResponseDto = kUserLoginResponseConverter.convert(kUser);
+
     loginResponseDto.setPhoto(user.get().getPhoto());
     loginResponseDto.setMembership(userBusiness);
+    loginResponseDto.setDetail(user.get());
     return loginResponseDto;
   }
 
@@ -186,11 +251,7 @@ public class UserServiceImpl implements UserService {
       throw new InternalServerException(ErrorMessage.INVALID_PHONE_NUMBER);
     }
     OtpDto otpDto = otpService.generateOtp(username, 180, messageText);
-    messageText = String.format(messageText, otpDto.getOtp());
-    SmsResponse response = smsService.sendSMS(username, messageText).block();
-    if (response == null) {
-      throw new InternalServerException(ErrorMessage.SMS_SENDING_ERROR);
-    }
+    notificationService.performSMSNotification(username, "", messageText);
     log.info("СМС с проверечным кодом был отправлен.Текст SMS : " + messageText);
   }
 
@@ -212,23 +273,32 @@ public class UserServiceImpl implements UserService {
         throw new NotFoundException(ErrorMessage.USER_NOT_FOUND_EXCEPTION);
       }
       kUser.setRequiredActions(new KAction[]{KAction.UPDATE_PASSWORD});
-      keycloakService.updateUser(kUser,token);
-      otpService.checkAndDeleteOtp(username,otp);
+      keycloakService.updateUser(kUser, token);
+      otpService.checkAndDeleteOtp(username, otp);
     }
     return checked;
   }
 
   @Override
-  @Transactional
-  public void deleteTemporalUser() {
+  @Transactional(rollbackFor = {InvalidRequestException.class, InternalServerException.class})
+  public void deleteTemporalUser(UserRole role) {
     TokenResponse token = keycloakService.getClientCredentialToken().block();
     if (token == null) {
       throw new NotAuthorizedException(ErrorMessage.INCORRECT_PASSWORD);
     }
-    usersRepository.findAllByTemporalIsTrue().forEach(userDetail -> {
-      keycloakService.deleteUser(userDetail.getId().toString(), token);
-      usersRepository.delete(userDetail);
+
+    List<UserDetail> toDelete = usersRepository.findAllByTemporalIs(Boolean.TRUE);
+    toDelete.forEach(userDetail -> {
+      if ((userBusinessRepository.findAllByUserUsername(userDetail.getUsername()).stream()
+          .anyMatch(userBusiness -> userBusiness.getUserRoles() == role)
+          || userBusinessRepository.findAllByUserUsername(userDetail.getUsername()).isEmpty()) &&
+          keycloakService.getUsers(token).stream()
+              .anyMatch(user -> user.getUsername().equals(userDetail.getUsername()))) {
+        keycloakService.deleteUser(userDetail.getId().toString(), token);
+        usersRepository.delete(userDetail);
+      }
     });
+
   }
 
   @Override
@@ -245,9 +315,10 @@ public class UserServiceImpl implements UserService {
     } else {
       keycloakService.setCredentials(user.get().getId().toString(), password, token);
       KUser kUser = KUser.builder().id(user.get().getId().toString())
-          .requiredActions(new KAction[]{KAction.UPDATE_PASSWORD, KAction.CONFIGURE_TOTP}).totp(false).enabled(false).build();
+          .requiredActions(new KAction[]{KAction.UPDATE_PASSWORD, KAction.CONFIGURE_TOTP})
+          .totp(false).enabled(false).build();
       keycloakService.updateUser(kUser, token);
-      keycloakService.logout(user.get().toString());
+      keycloakService.logout(user.get().getId().toString());
     }
   }
